third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-base-element/base-data-expected.txt

This is a testharness.js-based test.
[FAIL] Dynamically inserted first <base> has a data: URL so fallback is used
  assert_equals: expected "http://web-platform.test:8001/html/semantics/document-metadata/the-base-element/blah" but got "https://example.com/blah"
Harness: the test ran to completion.
